requirements.txt
pyjwt
fastapi
uvicorn
python-dotenv
pymongo
requests
bcrypt
pytz
pydantic[email]